backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
pyodbc
python-jose[cryptography]
passlib[bcrypt]
pydantic
openpyxl
reportlab
python-docx
